staff

+++ solar-panel-emulator/src/main/scala/ua/ucu/edu/model/SensorRecord.scala
@@ -5,4 +5,4 @@ import ua.ucu.edu.model.Location
 /**
   * To be used as a message in device topic
   */
-case class SensorRecord(panelId: String, location: Location, sensorType: String, measurement: Long)
+case class SensorRecord(panelId: String, location: Location, sensorType: String, measurement: Double)
